Оформил странички используя Bootstrap!

+++ src/main/java/org/makarovevg/mysbs/mysbs_app/controllers/AdminController.java
@@ -3,13 +3,19 @@ package org.makarovevg.mysbs.mysbs_app.controllers;
 
 import org.makarovevg.mysbs.mysbs_app.models.Person;
 import org.makarovevg.mysbs.mysbs_app.models.Role;
+import org.makarovevg.mysbs.mysbs_app.security.PersonDetails;
 import org.makarovevg.mysbs.mysbs_app.service.PersonService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.access.prepost.PostAuthorize;
 import org.springframework.security.access.prepost.PreAuthorize;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContextHolder;
+import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
+import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.ModelMap;
 import org.springframework.web.bind.annotation.*;
+import org.springframework.web.servlet.view.RedirectView;
 
 import java.util.List;
 
@@ -35,39 +41,56 @@ public class AdminController {
             personService.create(newPerson);
             model.addAttribute("newPerson", newPerson); // для таймлифа
         } catch (RuntimeException e) {
-            System.err.println(e.getMessage());
             return "admin/personAlreadyCreated";
         }
-        return "admin/personCreated";
+        return "redirect:/admin";
     }
 
 
     @GetMapping
     public String listUsers(ModelMap model) {
+
+        //получаем объект userDetails, который прошёл аутентификацию и авторизацию
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+
+        //вытаскиваем из принципал (объект успешной аутентификации, содержащий пользователя) нашего персона
+        PersonDetails personAuth = (PersonDetails) authentication.getPrincipal();
+
         List<Person> listPersons = personService.readPersons();
-        if (listPersons.isEmpty()) {
-            return "admin/emptyList";
+
+        for (Person p : listPersons) {
+            p.setPassword(p.getOriginalPassword());
         }
+
         model.addAttribute("listPersons", listPersons);
+        model.addAttribute("personAuth", personAuth.getPerson());
+        model.addAttribute("personInfo", personAuth.getPerson());
 
         return "admin/listPersons";
     }
 
     @PostMapping("/delete")
-    public String deleteUser(@RequestParam(value = "id", required = false) long id, ModelMap model) {
-        personService.delete(id);
-        model.addAttribute("personDeleted", id);
+    public String deleteUser(@RequestParam(value = "id", required = false) long id) {
 
-        return "/admin/personDeleted";
+        //получаем объект userDetails, который прошёл аутентификацию и авторизацию
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+
+        //вытаскиваем из принципал (объект успешной аутентификации, содержащий пользователя) нашего персона
+        PersonDetails personAuth = (PersonDetails) authentication.getPrincipal();
+
+        if (personAuth.getPerson().getId() == id) {
+            return "admin/personErrorDeleted";
+        }
+        personService.delete(id);
+        return "redirect:/admin";
     }
 
+
     @PostMapping("/update")
-    public String updateUser(@ModelAttribute Person person, ModelMap model) {
+    public String updateUser(@ModelAttribute Person person) {
 
         personService.update(person.getId(), person);
-        model.addAttribute("personUpdated", person.getId());
-
-        return "/admin/personUpdated";
+        return "redirect:/admin";
     }
 
 }
